### 2024-08-07 #7084 @editorialbot
[PRE REVIEW]: MDPSolver: An Efficient Solver for Markov Decision Processes
**Submitting author:** @areenberg (<a href="http://orcid.org/0000-0001-7146-3086">Anders Reenberg Andersen</a>)
**Repository:** https://github.com/areenberg/MDPSolver
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.9.7
