### 2024-09-23 #7084 @kellyrowland
hello @seankmartin @patrickmineault 👋 would you be interested in and available for reviewing this JOSS submission?
