### 2024-10-30 #7084 @kellyrowland
hello @virgesmith @davidtourigny 👋 would you be interested in and available for reviewing this JOSS submission?
