### 2024-10-30 #7084 @davidtourigny
HelloThanks for reaching out. Unfortunately, this is a bit out of scope for me and I don’t necessarily have the required expertise. Sorry about that.Best wishesDavidOn Oct 30, 2024, at 4:37 PM, Kelly L. Rowland ***@***.***> wrote:﻿
hello @virgesmith @davidtourigny 👋 would you be interested in and available for reviewing this JOSS submission?
