### 2024-11-03 #7084 @areenberg
Hi @kellyrowland,

I have a few suggestions:
- @williamlwj
- @mwetz
- @spinjet
- @GurjeetSinghSangra

All of the above have recently authored JOSS papers on somewhat related topics.

Best regards, Anders
