### 2024-11-25 #7084 @kellyrowland
hello @seankmartin @patrickmineault 👋 would you be interested in and available for reviewing this JOSS submission?
