### 2024-12-10 #7544 @MichaelHoltonPrice
Thank you for thinking of me. I am overcommitted right now, so with
apologies I don't think I can review it.

On Tue, Dec 10, 2024 at 9:45 AM Kelly L. Rowland ***@***.***>
wrote:

> hello @larryshamalama <https://github.com/larryshamalama>
> @MichaelHoltonPrice <https://github.com/MichaelHoltonPrice> 👋 would you
> be interested in and available for reviewing this JOSS submission?
>
