### 2025-01-13 #7544 @kellyrowland
hello @lwu9 @thomaspinder 👋 would you be interested in and available for reviewing this JOSS submission?
