### 2025-02-10 areenberg/MDPSolver#4 @gartavanis
JOSS Review
Dear authors,

Thanks for the work on this package and for submitting a paper to JOSS! Overall, the paper is well-written and it offers many useful features.

Below are a few comments that should be addressed before publication. There are no major issues, just small improvements.

Code-related comments:
- [x] Installation: When I tried to install, I also run into the glibc issue that has already been raised here (https://github.com/areenberg/MDPSolver/issues/3). It would be helpful to incorporate those instructions in the documentation under 'How to install'. That will be the first place where users might look for guidance, rather than the github issues.
- [x] In addition, you might want to consider creating containers with your package. In my experience, providing containers has been very helpful for users who are less familiar with installing/managing packages. (Just to be clear: creating containers is just a suggestion and not a requirement for the publication.)
- [x] Testing: There are no reference outputs to compare against. The tests run, but we can't be sure that they give the expected results. At the very least, you should add reference outputs and assert statements (or other mechanism that will make sure the test fails if the test results are incorrect). You should also add instructions on how to run the tests manually (perhaps in a dedicated section in the docs).
- [x] Contributing: You should add a short section in the documentation with contributing guidelines for external users who might want to add features, raise issues or ask for clarifications/support.

Manuscript-related comments:
- [x] In the manuscript you mention that you use a "parallelized efficient implementation of algorithms". It would be helpful to expand on this. For example: Does your package use multi-threading? Can you do multi-core runs? Does it take advantage of some parallelization technology like OpenMP? Just a sentence or two is enough.
- [x] "Our numerical experiments show that our implementation results in a substantially shorter runtime.": It would be appropriate to include a reference to your performance tests in the documentation.
- [x] Regarding the performance tests, the speedup for the Value Iteration Algorithm at 99% sparsity with 2500 states stands out from the rest in that graph. How do you explain this discrepancy?

Many thanks for addressing these comments!
