+++ README.md
@@ -23,7 +23,11 @@ Install directly from PyPI with:
 pip install mdpsolver
 ```
 
-MDPSolver works *out of the box* on Linux.
+MDPSolver works *out of the box* on Ubuntu 22 and newer.
+
+### GLIBC not found
+
+Some users will encounter the `version 'GLIBC_2.32' not found` error when attempting to import MDPSolver in Python. In this case, it might help to manually compile and replace the SO-file for the optimization module in the MDPSolver package. See the steps on how to solve the issue in the [documentation]([/guides/content/editing-an-existing-page](https://github.com/areenberg/MDPSolver/wiki#how-to-install)).
 
 ## Windows
 
