### 2025-02-15 9960c99
Reference output added to tests
