+++ Paper/paper.md
@@ -31,7 +31,7 @@ MDPSolver is a Python package for easy and fast optimization of Markov Decision 
 
 The industry currently experiences an increasing availability of data and computational resources. This development has created an opportunity for optimized decision-making based on mathematical modeling. Conversely, computational implementations can be costly, opening the door to software packages containing efficient and easily accessible tools. MDPSolver addresses this need by providing the accessibility of algorithms, optimality criteria, and other configurations. MDPSolver is highly applicable as a research tool since users can quickly employ and test various models and solution approaches. Due to the straightforward applicability of MDP models in decision-making sciences, such as operations research, MDPSolver has an important role in achieving any goals related to optimized decision-making under uncertainty.
 
-Several similar software packages are available for Python [@pymdptoolbox:2015], R [@MDPtoolbox:2017], Matlab [@Matlabtoolbox:2015], and Julia [@pomdps:2017] (see the section on related software below). MDPSolver targets Python and improves on the *pymdptoolbox* package [@pymdptoolbox:2015] by providing a parallelized efficient implementation of algorithms from MDP theory [@Puterman:1994a]. Our numerical experiments show that our implementation results in a substantially shorter runtime. Moreover, MDPSolver provides the user with the option to experiment with various value update methods that are not available in the *pymdptoolbox* package. In an upcoming version, MDPSolver will also contain a selection of built-in MDP problems often considered in the scientific literature, such as multi-component replacement, job scheduling, and inventory management. The built-in problems will enable users to evaluate even larger models by computing the model parameters (e.g. transition probabilities) on demand.
+Several similar software packages are available for Python [@pymdptoolbox:2015], R [@MDPtoolbox:2017], Matlab [@Matlabtoolbox:2015], and Julia [@pomdps:2017] (see the section on related software below). MDPSolver targets Python and improves on the *pymdptoolbox* package [@pymdptoolbox:2015] by providing a parallelized efficient implementation of algorithms from MDP theory [@Puterman:1994a]. Specifically, MDPSolver uses OpenMP [openMP:2025], ensuring that large vector validations are distributed across multiple threads and executed on multiple cores in parallel. Our numerical experiments show that our implementation results in a substantially shorter runtime [wikiPerformance:2025]. Moreover, MDPSolver provides the user with the option to experiment with various value update methods that are not available in the *pymdptoolbox* package. In an upcoming version, MDPSolver will also contain a selection of built-in MDP problems often considered in the scientific literature, such as multi-component replacement, job scheduling, and inventory management. The built-in problems will enable users to evaluate even larger models by computing the model parameters (e.g. transition probabilities) on demand.
 
 In our past research, @Andersen:2022b and @Andersen:2022a used earlier versions of MDPSolver for conducting numerical experiments on multi-component replacement problems. In the future, we are planning on utilizing MDPSolver for conducting experiments in our research on the optimization of hospital patient flow.
 
