+++ Paper/paper.bib
@@ -59,6 +59,20 @@
 	Publisher={MathWorks}
 }
 
+@Manual{openMP:2025,
+    Title = {The OpenMP API specification for parallel programming},
+    url = {https://www.openmp.org/},
+    urldate={2025-02-15},
+    Publisher={The OpenMP Architecture Review Board}
+}
+
+@Manual{wikiPerformance:2025,
+    Title = {MDPSolver wiki: Performance experiments},
+    url = {https://github.com/areenberg/MDPSolver/wiki/Performance-experiments},
+    urldate={2025-02-15},
+    Publisher={MDPSolver Github}
+}
+
 @article{pomdps:2017,
   Author = {{Egorov}, M. and {Sunberg}, Z.~N. and {Balaban}, E. and {Wheeler} T.~A. and {Gupta}, J.~K. and {Kochenderfer}, M.~J.},
   Title   = {{POMDP}s.jl: A Framework for Sequential Decision Making under Uncertainty},
