### 2025-02-15 areenberg/MDPSolver#4 @areenberg
Dear @gartavanis,

Thank you for reviewing our manuscript and for your constructive comments. We have addressed your comments on the code, documentation, and manuscript. Please note that our response to each comment is highlighted in italics below.

**Code-related comments**

Installation: When I tried to install, I also run into the glibc issue that has already been raised here (https://github.com/areenberg/MDPSolver/issues/3). It would be helpful to incorporate those instructions in the documentation under 'How to install'. That will be the first place where users might look for guidance, rather than the github issues.
*Following your recommendations, the section on 'How to install' in the documentation now contains a description of how to fix the glibc issue by manually compiling the SO-file.*

In addition, you might want to consider creating containers with your package. In my experience, providing containers has been very helpful for users who are less familiar with installing/managing packages. (Just to be clear: creating containers is just a suggestion and not a requirement for the publication.)
*Containers are already on our drawing board and will surely be added to the repository in the near future.*

Testing: There are no reference outputs to compare against. The tests run, but we can't be sure that they give the expected results. At the very least, you should add reference outputs and assert statements (or other mechanism that will make sure the test fails if the test results are incorrect). You should also add instructions on how to run the tests manually (perhaps in a dedicated section in the docs).
*The tests now contain reference values as suggested. The tests will automatically compare the generated output to the reference values and abort the program if any discrepancies are detected. The user will be notified if a test fails, and correspondingly if all tests were successful. Moreover, the documentation now includes guidelines on how to run the tests manually.*

Contributing: You should add a short section in the documentation with contributing guidelines for external users who might want to add features, raise issues or ask for clarifications/support.
*A brief 'Contribution guidelines' section has been added to the documentation showing how to add features, raise issues, and contact the authors.*

**Manuscript-related comments**

In the manuscript you mention that you use a "parallelized efficient implementation of algorithms". It would be helpful to expand on this. For example: Does your package use multi-threading? Can you do multi-core runs? Does it take advantage of some parallelization technology like OpenMP? Just a sentence or two is enough.
*We have added a brief description of the technology used for parallel computing (OpenMP) and clarified that the package employs multi-threading executed on multiple cores.*

"Our numerical experiments show that our implementation results in a substantially shorter runtime.": It would be appropriate to include a reference to your performance tests in the documentation.
*A reference to our performance tests in the documentation has been added to the manuscript.*

Regarding the performance tests, the speedup for the Value Iteration Algorithm at 99% sparsity with 2500 states stands out from the rest in that graph. How do you explain this discrepancy?
*A careful examination of the test parameters and absolute runtimes did not reveal any obvious explanation for the discrepancy. We deem this was caused by pymdptoolbox's implementation adapting to the problem parameters (e.g. the state space size). We can share the details of the results upon request.*
