### 2025-02-15 26bd7ba
Fixed paper compilation issue
