### 2025-02-16 areenberg/MDPSolver#5 @braniii
WIP: JOSS Review 2
Hi @areenberg this is the issue to track my review, https://github.com/openjournals/joss-reviews/issues/7544.

I should note that I am using Manjaro Linux. For that, I can not follow the `apt-get` commands, and I try to follow as close as possible to the instructions

## Issues and Questions:

**Don't feel forced to change things if you disagree. You can also simply comment briefly on the points.**

- [x] [compiling-the-c-module](https://github.com/areenberg/MDPSolver/wiki/Contribution-guidelines#compiling-the-c-module):
  > "Install `python3.10-config` by running `sudo apt-get install python3.10-dev`"

  This raises two questions:

  1. As you've stated, MDPSolver works with py3.13. Is there a need to use a version which is eol?
  2. This suggests the users to install Python, pybind, and MDPSolver globally. In many modern Linux distros, e.g. all Arch-based ones, this is prevented. I guess it would be beneficial to give the commands to create a minimal virtual env, either using `conda`, `venv`, `uv`, etc, to unify the setup among all linux distros.
- [x] [compiling-the-c-module](https://github.com/areenberg/MDPSolver/wiki/Contribution-guidelines#compiling-the-c-module): I could follow the Linux instructions step 1-4. However, I am missing the part on how to install the modified package. Using the suggested `pip install mdpsolver` installs the uploaded version.
- [x] Instead of letting the user compile everything by hand, I think it would make a contribution from the community much more likely if you set up a setup.py an pyproject.toml with pybind11 so that you can install the package locally in editable mode via: `python -m pip install --editable .` Here is an example on what I mean: https://github.com/pybind/python_example
- [x] Would you mind setting up a Github action (CI) which automatically builds the package and runs the tests?
- [x] Your Python code does not follow the PEP8 style guide. Would you mind to format it with, e.g. black?
- [x] Judging from the source code, it looks like cbmaux and tbmaux are available but not yet documented.
- [x] Using modern IDEs, the docstring is parsed and shown to users when coding. Therefore, I would strongly encourage you to write (at least) short docstrings in your `model.py`.
- [x] Where can I find the script used to generate the [benchmark section](https://github.com/areenberg/MDPSolver/wiki/Performance-experiments)?

## Optional suggestions for improvement:

- [x] Is there a reason why you are not using one of the common Python test suits, e.g. pytest? This would also allow to show how much of your (Python) API is covered by the tests.
- [x] To make usage of github, it might be helpful to copy the "contribution guidelines" to the github `CONTRIBUTING.md` file.
- [x] Is there a reason why the `CMakeLists.txt` is not included in the repository, see [compiling-the-c-module](https://github.com/areenberg/MDPSolver/wiki/Contribution-guidelines#compiling-the-c-module)?
