### 2025-02-17 #7544 @gartavanis
@kellyrowland The issues I raised have been addressed satisfactorily by the authors. I think the paper is well-written and it offers many useful features. The software meets all the requirements to be accepted by JOSS and I suggest it be accepted as is.
