### 2025-03-22 588a2e8
Python code formatted and docstrings added
