+++ Benchmark/runtime_tests_pi.py
@@ -9,9 +9,10 @@ from scipy.sparse import coo_matrix
 from scipy.sparse import csr_matrix
 
 
-#--------------------------------------------
+# --------------------------------------------
 #   FUNCTIONS
-#--------------------------------------------
+# --------------------------------------------
+
 
 def sampleProbs(n):
     rnds = [random.expovariate(1) for _ in range(n)]
@@ -19,112 +20,157 @@ def sampleProbs(n):
     rnds_norm = [x / sm for x in rnds]
     return rnds_norm
 
+
 def randomMDP(nStates, nActions, sparsity):
-    
+
     nJumps = math.ceil(nStates * (1 - sparsity))
-    
-    #mdpsolver format
+
+    # mdpsolver format
     tranMatProbs_mdpsolver = [[[] for _ in range(nActions)] for _ in range(nStates)]
     tranMatColumns_mdpsolver = [[[] for _ in range(nActions)] for _ in range(nStates)]
-    rewards_mdpsolver = [[0 for _ in range(nActions)] for _ in range(nStates)] 
-        
-    rndrew = 0    
+    rewards_mdpsolver = [[0 for _ in range(nActions)] for _ in range(nStates)]
+
+    rndrew = 0
     for sidx in range(nStates):
         for aidx in range(nActions):
-            rndrew = random.gauss(rndrew,1)
+            rndrew = random.gauss(rndrew, 1)
             rewards_mdpsolver[sidx][aidx] = rndrew
 
             probs = sampleProbs(nJumps)
             cols = random.sample(range(nStates), nJumps)
             tranMatProbs_mdpsolver[sidx][aidx] = probs
             tranMatColumns_mdpsolver[sidx][aidx] = sorted(cols)
-    
-    #mdptoolbox format
+
+    # mdptoolbox format
     transitions_toolbox = np.empty(nActions, dtype=object)
     rewards_toolbox = np.zeros((nStates, nActions))
     for aidx in range(nActions):
-        vals = np.zeros(nStates*nJumps)
-        rowIndices = np.zeros(nStates*nJumps)
-        colIndices = np.zeros(nStates*nJumps)
-        k=0
+        vals = np.zeros(nStates * nJumps)
+        rowIndices = np.zeros(nStates * nJumps)
+        colIndices = np.zeros(nStates * nJumps)
+        k = 0
         for sidx in range(nStates):
             rewards_toolbox[sidx][aidx] = rewards_mdpsolver[sidx][aidx]
             for jidx in range(nJumps):
                 vals[k] = tranMatProbs_mdpsolver[sidx][aidx][jidx]
                 rowIndices[k] = sidx
                 colIndices[k] = tranMatColumns_mdpsolver[sidx][aidx][jidx]
-                k+=1
-        mat_coo = coo_matrix((vals,(rowIndices,colIndices)),shape=(nStates,nStates))
+                k += 1
+        mat_coo = coo_matrix((vals, (rowIndices, colIndices)), shape=(nStates, nStates))
         mat_csr = mat_coo.tocsr()
-        transitions_toolbox[aidx]=mat_csr
-        
-    return rewards_mdpsolver, tranMatProbs_mdpsolver, tranMatColumns_mdpsolver, rewards_toolbox, transitions_toolbox
+        transitions_toolbox[aidx] = mat_csr
+
+    return (
+        rewards_mdpsolver,
+        tranMatProbs_mdpsolver,
+        tranMatColumns_mdpsolver,
+        rewards_toolbox,
+        transitions_toolbox,
+    )
 
 
-#--------------------------------------------
+# --------------------------------------------
 #   TESTS
-#--------------------------------------------
+# --------------------------------------------
 
-#discount
+# discount
 discount = 0.99
-#tolerance
[…]

[…]
