### 2025-03-22 areenberg/MDPSolver#5 @areenberg
Hi @braniii, thank you so much for reviewing our manuscript! Please find our responses to your comments below (note that our responses are highlighted in *italics*):

compiling-the-c-module: "Install python3.10-config by running sudo apt-get install python3.10-dev". This raises two questions:

1. As you've stated, MDPSolver works with py3.13. Is there a need to use a version which is eol?
*Thank you for this comment. As a result of your comment, we have updated the documentation using Python 3.13 as a basis for the guidelines.*

2. This suggests the users to install Python, pybind, and MDPSolver globally. In many modern Linux distros, e.g. all Arch-based ones, this is prevented. I guess it would be beneficial to give the commands to create a minimal virtual env, either using conda, venv, uv, etc, to unify the setup among all linux distros.
*We have followed your recommendations and added a step showing how to create a minimal virtual environment using conda.*

* compiling-the-c-module: I could follow the Linux instructions step 1-4. However, I am missing the part on how to install the modified package. Using the suggested `pip install mdpsolver` installs the uploaded version.
*Thank you for noticing. We have added a new sub-section in the documentation showing how to install the modified module in the MDPSolver package.*

* Instead of letting the user compile everything by hand, I think it would make a contribution from the community much more likely if you set up a setup.py an pyproject.toml with pybind11 so that you can install the package locally in editable mode via: `python -m pip install --editable`. Here is an example on what I mean: https://github.com/pybind/python_example.
*Thank you for the suggestion. While we agree that establishing a procedure for automatically compiling and installing the module would encourage more contributions, this setup is beyond the scope of the current development work. We hope you will be satisfied with this being implemented in the near future.*

* Would you mind setting up a Github action (CI) which automatically builds the package and runs the tests?
*Like the above, we agree that establishing such a procedure would encourage more contributions and enhance transparency. However, this setup is beyond the scope of the current development work. We hope you are satisfied with the current setup, despite it being more manual.*

* Your Python code does not follow the PEP8 style guide. Would you mind to format it with, e.g. black?
*We have followed your suggestion and formatted all PY-files with black.*

* Judging from the source code, it looks like cbmaux and tbmaux are available but not yet documented.
*The cbmaux and tbmaux files were not fully implemented and tested, which is why they were not documented. We have now removed them from the repository to avoid further confusion.*

* Using modern IDEs, the docstring is parsed and shown to users when coding. Therefore, I would strongly encourage you to write (at least) short docstrings in your model.py.
*We have followed your suggestion and added docstrings to all methods in the model.py file.*

* Where can I find the script used to generate the benchmark section?
*We have added a folder named ‘Benchmark’ that contains the files for the benchmark tests.*

**Optional suggestions for improvement:**

* Is there a reason why you are not using one of the common Python test suits, e.g. pytest? This would also allow to show how much of your (Python) API is covered by the tests.
*We recognize the value of using test suites like pytest for ensuring comprehensive test coverage. However, due to the current development focus, we have opted for a more manual testing process.*

* To make usage of github, it might be helpful to copy the "contribution guidelines" to the github CONTRIBUTING.md file.
*We have decided to keep the documentation for MDPSolver as consolidated as possible. Thus, users can find most of the fundamental information in the wiki.*

* Is there a reason why the CMakeLists.txt is not included in the repository, see compiling-the-c-module?
*We have decided to keep the source code as "clean" as possible. The CMakeLists.txt file would only be relevant to Windows developers, therefore, we have left it out of the repository.*
