### 2025-03-31 areenberg/MDPSolver#5 @braniii
@areenberg Thank you for addressing all points. I will try to go through your answers the next days.

- [x] I was able to manually compile the CPP module now. But since the package cannot be installed from the local folder (e.g. with `python -m pip install -e .`), I was not able to run the tests. I get the following error
`ModuleNotFoundError: No module named 'mdpsolver'`
