### 2025-04-16 6fd90f1
Automatically build and test the compiled optimization module
