+++ Python/tests/test1.py
@@ -1,9 +1,12 @@
-import mdpsolver
 import random
 import sys
+import os
 import numpy as np
 from random import randint
-sys.path.append("Python/src/")
+project_root = os.path.dirname(os.path.abspath(__file__))
+src_path = os.path.join(project_root, "..", "src")
+sys.path.append(os.path.abspath(src_path))
+import mdpsolver
 
 # TEST 1
 # Simple MDP with 3 states and 2 actions in each state.
