+++ Python/tests/test2.py
@@ -1,8 +1,13 @@
-import mdpsolver
 import example_model
+import random
 import sys
+import os
 import numpy as np
-sys.path.append("Python/src/")
+from random import randint
+project_root = os.path.dirname(os.path.abspath(__file__))
+src_path = os.path.join(project_root, "..", "src")
+sys.path.append(os.path.abspath(src_path))
+import mdpsolver
 
 # TEST 2
 # In this test, we load the rewards and transition probabilities
