+++ Python/tests/test2.py
@@ -1,9 +1,7 @@
 import example_model
-import random
 import sys
 import os
 import numpy as np
-from random import randint
 project_root = os.path.dirname(os.path.abspath(__file__))
 src_path = os.path.join(project_root, "..", "src")
 sys.path.append(os.path.abspath(src_path))
