### 2025-04-16 3af8182
Added test files for the pip installation of MDPSolver
