+++ Python/tests/test1_pip.py
@@ -1,0 +1,713 @@
+import random
+import sys
+import numpy as np
+from random import randint
+import mdpsolver
+
+# TEST 1
+# Simple MDP with 3 states and 2 actions in each state.
+
+# ---------------------------------------
+# CONFIGURATION 1
+# ---------------------------------------
+
+# rewards
+# 1st index: from (current) states
+# 2nd index: actions
+rewards = [[5, -1], [1, -2], [50, 0]]
+
+# transition probabilities
+# 1st index: from (current) states
+# 2nd index: actions
+# 3rd index: to (next) states
+tranMatWithZeros = [
+    [[0.9, 0.1, 0.0], [0.1, 0.9, 0.0]],
+    [[0.4, 0.5, 0.1], [0.3, 0.5, 0.2]],
+    [[0.2, 0.2, 0.6], [0.5, 0.5, 0.0]],
+]
+
+# initial policy
+random.seed(10)
+initPolicy = [randint(0, 1) for p in range(0, 3)]
+
+# Model 1a (discounted reward, parallel)
+mdl1a = mdpsolver.model()
+mdl1a.mdp(discount=0.95, rewards=rewards, tranMatWithZeros=tranMatWithZeros)
+mdl1a.solve(algorithm="mpi", update="standard", parallel=True, initPolicy=initPolicy)
+if not np.array_equal(np.array(mdl1a.getPolicy()), np.array([1, 1, 0])):
+    sys.exit("Model 1a failed!")
+if not np.array_equal(
+    np.round(np.array(mdl1a.getValueVector()), 3),
+    np.round(np.array([200.00114718191236, 212.86672887958622, 298.7090458676583]), 3),
+):
+    sys.exit("Model 1a failed!")
+
+# Model 2a (discounted reward, parallel)
+mdl2a = mdpsolver.model()
+mdl2a.mdp(discount=0.95, rewards=rewards, tranMatWithZeros=tranMatWithZeros)
+mdl2a.solve(algorithm="pi", update="standard", parallel=True, initPolicy=initPolicy)
+if not np.array_equal(np.array(mdl2a.getPolicy()), np.array([1, 1, 0])):
+    sys.exit("Model 2a failed!")
+if not np.array_equal(
+    np.round(np.array(mdl2a.getValueVector()), 3),
+    np.round(np.array([200.00114718191236, 212.86672887958622, 298.7090458676583]), 3),
+):
+    sys.exit("Model 2a failed!")
+
+# Model 3a (discounted reward, parallel)
+mdl3a = mdpsolver.model()
+mdl3a.mdp(discount=0.95, rewards=rewards, tranMatWithZeros=tranMatWithZeros)
+mdl3a.solve(algorithm="vi", update="standard", parallel=True)
+if not np.array_equal(np.array(mdl3a.getPolicy()), np.array([1, 1, 0])):
+    sys.exit("Model 3a failed!")
+if not np.array_equal(
+    np.round(np.array(mdl3a.getValueVector()), 3),
+    np.round(np.array([200.0013199959708, 212.86689665815788, 298.709197760198]), 3),
+):
+    sys.exit("Model 3a failed!")
+
+# Model 1b (discounted reward, unparallel)
+mdl1b = mdpsolver.model()
+mdl1b.mdp(discount=0.95, rewards=rewards, tranMatWithZeros=tranMatWithZeros)
+mdl1b.solve(algorithm="mpi", update="standard", parallel=False, initPolicy=initPolicy)
+if not np.array_equal(np.array(mdl1b.getPolicy()), np.array([1, 1, 0])):
+    sys.exit("Model 1b failed!")
[…]
