### 2025-04-16 areenberg/MDPSolver#5 @areenberg
Hi @braniii, thank you for your comment.
I have updated the setup of the test files (i.e., `test1.py` and `test2.py`) so they now import the MDPSolver package from the `Python/src` directory. You should be able to run both files without any issues.

Additionally, while addressing your comment, I created a shell script to automatically compile, install, and test the updated module. I’ve also created a corresponding BAT file for Windows users. Note that the *contribution guidelines* have been updated accordingly.

Let me know if you have any further comments or questions :)
