### 2025-05-04 areenberg/MDPSolver#5 @braniii
@areenberg thx for your work. I've successfully used it to recompile and test the package locally. I was also able to rerun the benchmarks. Now that you have the bash script, most work is done to setup CI testing.

As a final comment: I've checked your CPP code and it would benefit from more comments.
