### 2025-05-04 #7544 @braniii
@kellyrowland Sorry for the delay. The author has addressed all the major issues and I finished my review.

The Paper is well structured. However, I would like to make one more comment: For a Python package, I would say that the community standards provide that the project included a `setup.py/.toml` to easily work on it and that there are automated tests via github CI. Neither is the case with this project. But you also have to say that this project implements everything in cpp and only has Python bindings. Therefore, I would assume that you have to compare it with the community standards of cpp, which it meets.

Therefore, I suggest that it is accepted as is.
