+++ Paper/paper.bib
@@ -40,7 +40,8 @@
     Year = 2017,
     Note = {R package version 4.0.3},
     url = {https://CRAN.R-project.org/package=MDPtoolbox},
-	Publisher={The R Foundation}
+    Publisher={The R Foundation},
+    doi={10.32614/cran.package.mdptoolbox}
 }
 
 @Manual{Matlabtoolbox:2015,
