### 2025-05-10 #7544 @areenberg
Hi @kellyrowland,

We’re happy to hear that the paper has been accepted! Please find our responses to each step of the wrap-up process below:

* We have added the DOI ([10.32614/cran.package.mdptoolbox](10.32614/cran.package.mdptoolbox)) to 'MDPtoolbox: Markov Decision Processes Toolbox' and checked that the paper compiles correctly.
* This is the link for the latest release of our software, which has been updated according to the comments from the reviewers: https://github.com/areenberg/MDPSolver/releases/tag/v0.9.9
* The release has been archived on Zenodo with the following DOI: [10.5281/zenodo.15069051](https://doi.org/10.5281/zenodo.15069051)

Please let me know if there’s anything else I can assist with.
