### 2025-05-15 #7544 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
